% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\docType{data}
\name{abbreviations}
\alias{abbreviations}
\alias{abbreviations}
\title{Abbreviations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 93 rows and 2 columns.}
\usage{
abbreviations

abbreviations
}
\description{
A short list of abbreviations found in the headlines and their expansion.

A list of regular expression patterns that typically indicate names of people
or organizations found in the headlines.
}
\details{
\describe{

\item{abbreviation}{A regular expression capturing an abbreviation
found in the headlines.}

\item{replacement}{An expansion of the abbreviation. Used in conjunction with
\code{\link[qdap:mgsub]{qdap::mgsub()}} to expand the abbreviations in the headlines in the file
\link{processed_corpus}. See the \code{process_corpus} file in the \href{https://github.com/xmarquez/KCNA/data-raw}{data-raw folder onGithub} for details.}
}

\describe{

\item{pattern}{A regular expression capturing an abbreviation found in the
headlines.}

\item{change_to}{A replacement that expands, uppercases, or disambiguates the
regex match in the headlines. Used in conjunction with
\code{\link[qdap:mgsub]{qdap::mgsub()}} to expand the abbreviations in the headlines in the file
\link{processed_corpus}. See the \code{process_corpus} file in the \href{https://github.com/xmarquez/KCNA/data-raw}{data-raw folder onGithub} for details.}
}
}
\seealso{
Other other data: \code{\link{country_mentions}},
  \code{\link{person_categories}}

Other other data: \code{\link{country_mentions}},
  \code{\link{person_categories}}
}
\keyword{datasets}
