% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\docType{data}
\name{country_mentions}
\alias{country_mentions}
\title{Country mentions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 71742 rows and 5 columns.}
\usage{
country_mentions
}
\description{
A data frame containing country and demonym mentions in the headlines. It is
as comprehensive as possible, but there may be errors in it.
}
\details{
\describe{

\item{country_name}{The country name of the country or demonym found in the
headline.}

\item{demonym_match}{An expression capturing the demonym (e.g., American,
Russian) found in the headline, if anything. If a country name is found, but
not a demonym, this defaults to \code{NA}. }

\item{processed_title}{The title of the headline, as in \link{processed_corpus}.
(This will not match exactly the original headline due to abbreviation
expansion, lowercasing, and the like, but it can be used for joins with
\link{processed_corpus}).}

\item{lon}{The approximate longitude of the country.}

\item{lat}{The approximate latitude of the country.}

}
}
\seealso{
Other other data: \code{\link{abbreviations}},
  \code{\link{person_categories}}
}
\keyword{datasets}
