% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_sankey_network}
\alias{create_sankey_network}
\title{Create a Sankey Flow Diagram from a series of texts.}
\usage{
create_sankey_network(text, node, type = c("after", "before"), title = NULL,
  context_words = 4, num_levels = 15)
}
\arguments{
\item{text}{A vector of sentences used to generate concordances around a
central word. Though texts of any length can be used, the function
currently works best if the texts are sentences (like the KCNA headlines in
this package) rather than full paragraphs.}

\item{node}{The central word in the concordance (around which the flow of
text will be computed). This is a regular expression; so, for example,
\code{"[Nn]ation(al)?"} will compute concordances around \code{nation},
\code{national}, \code{Nation}, and \code{National}. And for various internal
reasons, \code{"friend"} will typically match \code{"friendship"} as well as
\code{"friend"}. To avoid this, surround the expression with \code{\\b}
(e.g., \code{"\\bfriend\\b"}) will match only \code{"friend"} and not
\code{"friendship"}).}

\item{type}{Whether to create the Sankey network "after" the word, or
"before" the word.}

\item{title}{A title for the visualization, if wanted. Defaults to NULL.}

\item{context_words}{The number of words of context for the concordance.
Defaults to 4.}

\item{num_levels}{The maximum number of terms to show in each column of the
graphical summary. Defaults to showing the most common 15 terms.}
}
\value{
A networkD3 htmlwidget. It's automatically included in knitr.
}
\description{
Create a Sankey Flow Diagram from a series of texts.
}
\examples{
\dontrun{
create_sankey_network(headlines$title, "Kim Jo[n|m]g Il")
create_sankey_network(headlines$title, "Kim Jo[n|m]g Il", type = "before")
create_sankey_network(headlines$title, "Kim Jo[n|m]g Il", context_words = 5,
num_levels = 20)
create_sankey_network(headlines$title, "Kim Jong Il|Kim Jong Un")
}
}
