% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{findpeaks}
\alias{findpeaks}
\title{Find local peaks in a vector}
\usage{
findpeaks(vec, bw = 1, x.coo = c(1:length(vec)))
}
\arguments{
\item{vec}{a vector}

\item{bw}{the bandwith for local peak search}

\item{x.coo}{the set of index values where peaks are to be located}
}
\value{
A list with local maxima and minima and their location in the
original vector. From
https://www.r-bloggers.com/an-algorithm-to-find-local-extrema-in-a-vector/,
by Timothee Poisot
}
\description{
Find local peaks in a vector
}
\examples{
x <- runif(100)
x
findpeaks(x, 2)
findpeaks(x, 25)
}
