% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\docType{data}
\name{headlines}
\alias{headlines}
\title{KCNA headlines, January 1997-September 2014}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 114641 rows and 6 columns.}
\usage{
headlines
}
\description{
A collection of all headlines in the Japanese-hosted \href{http://www.kcna.co.jp/index-e.htm}{KCNAwebsite} as of September 2014.
}
\details{
\describe{

\item{date}{Date of the headline. N = 114633, unique = 6087, min =
1997-01-01, max = 2014-09-25.}

\item{title}{The actual headline. N = 114633, unique = 101878. }

\item{english}{For \code{english_prob > 0.5}, we assume the headline is in
English. This is a reasonable assumption, but it may not be perfect; some
spanish language headlines may slip through, and some english headlines may
be misclassified, despite extensive checking.}

\item{english_prob}{A number signifying the degree of agreement of four
different algorithms with respect to the question of whether the headline is
in English. 0 means none of the algorithms used classified the headline as
being in English; 1 means all of them do. See the \code{language_detection} file
in the \href{https://github.com/xmarquez/KCNA/data-raw}{data-raw folder onGithub} for details. Headlines not
in english are in spanish.}

\item{story.link}{Link to the original story in the KCNA website. May have
changed since headlines were harvested; KCNA has been known to delete stories
or change links. Unique = 114319. }

\item{index.link}{Link to the original index page in the KCNA website. May
have changed since headlines were harvested; KCNA has been known to delete
stories or change links. Unique = 6087. }

}
}
\seealso{
Other headline data: \code{\link{preprocessed_corpus}},
  \code{\link{processed_corpus}}
}
\keyword{datasets}
