% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\docType{package}
\name{KCNA-package}
\alias{KCNA}
\alias{KCNA-package}
\title{Headlines from the Japanese-hosted \href{http://www.kcna.co.jp/index-e.htm}{KCNA website}
as of September 2014.}
\description{
Headlines from the Korean Central News agency, January 1997-September 2014,
in a format suitable for Natural Language Processing, along with several other
small datasets and convenience functions to work with these headlines. Also
includes most of the KCNA stories from 2011 and 2012.
}
\details{
The main dataset is \link{preprocessed_corpus}, though you may also need to use
\link{processed_corpus} for some tasks.
}
\author{
\strong{Maintainer}: Xavier Marquez \email{marquez.x@gmail.com}

}
