% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\docType{data}
\name{KCNA_2011}
\alias{KCNA_2011}
\title{KCNA stories, 2011}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13907 rows and 5 columns.}
\usage{
KCNA_2011
}
\description{
All KCNA stories for 2011 harvested from the Japanese-hosted \href{http://www.kcna.co.jp/index-e.htm}{KCNAwebsite} as of September 2014
}
\details{
\describe{ \item{story}{The actual story. N = 11997, unique = 11940. }

\item{date}{Date of the story. N = 13907, unique = 363, min = 2011-01-01, max
= 2011-12-31. }

\item{index.link}{Original link to the index page. N = 13907, unique = 363.}

\item{language}{Whether the story was in English ("en") or Spanish ("es"),
according to KCNA itself. This classification is not wholly accurate; some
Spanish language stories appear as English language and vice-versa. N =
13907, unique = 2. Best to use the classification in \link{headlines}. }

\item{title}{The headline. N = 13322, unique = 12689. }
}
}
\seealso{
Other story data: \code{\link{KCNA_2012}}
}
\keyword{datasets}
