% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\docType{data}
\name{KCNA_2012}
\alias{KCNA_2012}
\title{KCNA stories, 2012}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10869 rows and 5 columns.}
\usage{
KCNA_2012
}
\description{
All stories for January-November 2012 harvested from the Japanese-hosted
\href{http://www.kcna.co.jp/index-e.htm}{KCNA website} as of September 2014
}
\details{
\describe{

\item{story}{The actual story. N = 10340, unique = 10307.}

\item{date}{Date of the story. N = 10869, unique = 335, min = 2012-01-01, max
= 2012-11-30.}

\item{index.link}{Original link to the index page. May have changed since the
stories were harvested. N = 10869, unique = 10838. }

\item{language}{Whether the story was in English ("en") or Spanish ("es"),
according to KCNA itself. N = 10869, unique = 2. This classification is not
wholly accurate; some Spanish language stories appear as English language and
vice-versa. Best to use the classification in \link{headlines}.  }

\item{title}{The headline. N = 10869, unique = 10400.}

}
}
\seealso{
Other story data: \code{\link{KCNA_2011}}
}
\keyword{datasets}
