% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\docType{data}
\name{korean_names}
\alias{korean_names}
\title{Korean name components}
\format{An object of class \code{list} of length 3.}
\usage{
korean_names
}
\description{
A list with the three possible components of a Korean name. \href{https://en.wikipedia.org/wiki/List_of_Korean_given_names}{Harvested fromWikipedia} and
updated manually to take into account KCNA's spelling preferences and some
unique names. Used in the file \code{Extracting_North_Korean_politicians.Rmd} in
the \href{https://github.com/xmarquez/KCNA/data-raw}{data-raw folder on Github};
check that file for details.
}
\details{
\describe{

The list has three slots.

\item{family_name}{Possible family names. N = 1002.}

\item{first_syllable}{Possible first syllable of name. N = 268.}

\item{Second syllable}{Possible second syllable of name. N = 266.}

}
}
\keyword{datasets}
