% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\docType{data}
\name{person_categories}
\alias{person_categories}
\title{Person categories}
\format{An object of class \code{data.frame} with 86 rows and 4 columns.}
\usage{
person_categories
}
\description{
A data frame containing identifying material about the Korean people
mentioned in the headlines, taken from Wikipedia.
}
\details{
\describe{

\item{mentioned}{The person mentioned. Names are lowercased and formatted
with underscores instead of spaces.}

\item{category}{A broad category to which the person belongs, according to
Wikipedia. People can belong to more than one category; for example, Kim Jong
Il belongs to "Kim dynasty (North Korea)", "Heads of state of North Korea",
"Members of the Supreme People's Assembly" and Worker's Party of Korea
Politicians". Not every person mentioned in the headlines is found in
Wikipedia, or has sufficient information associated with it in the DBpedia
ontology to assign a category. This is especially the case for minor North
Korean politicians mentioned perhaps once or twice in the headlines.}

\item{uri}{The \url{http://dbpedia.org} link from which the
information was scraped.}

\item{n}{The number of mentions of the person in the English-language
headlines.}

}
}
\seealso{
Other other data: \code{\link{abbreviations}},
  \code{\link{country_mentions}}
}
\keyword{datasets}
