% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\docType{data}
\name{potential_names}
\alias{potential_names}
\title{Potential names}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 612 rows and 2 columns.}
\usage{
potential_names
}
\description{
A data frame with patterns that are likely to be named entities within the headlines.
}
\details{
\describe{

\item{pattern}{A regular expression indicating a named entity.}

\item{change_to}{A string indicating a useful substitution.}
}
}
\keyword{datasets}
