% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\docType{data}
\name{processed_corpus}
\alias{processed_corpus}
\title{A syntactically parsed corpus of the English-language KCNA headlines}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 766667 rows and 6 columns.}
\usage{
processed_corpus
}
\description{
A syntactically parsed version and cleaned up version of the English-language
headlines  of \link{preprocessed_corpus}, suitable for some text analysis tasks.
Includes part-of-speech (POS) tags for the words in the headlines, generated
by running the corpus through the \code{\link[openNLP:Maxent_POS_Tag_Annotator]{openNLP::Maxent_POS_Tag_Annotator()}} POS
annotator.
}
\details{
\describe{ \item{date}{Date of the headline. Unique = 6087, min = 1997-01-01,
max = 2014-09-25.}

\item{processed_title}{The pre-processed headline title, for use with the
\code{\link[openNLP:Maxent_POS_Tag_Annotator]{openNLP::Maxent_POS_Tag_Annotator()}} annotator. Processing the title
involves lowercasing, transforming many names to unigrams (e.g., "Kim Jong
Il" to "Kim_Jong_Il"), expanding many appreviations and standardazing
mentions of many organizations in preparation for part-of-speech (POS)
tagging. For details of the exact pre-processing steps, See the
\code{process_corpus} file in the \href{https://github.com/xmarquez/KCNA/data-raw}{data-raw folder onGithub}.}

\item{words}{The word tokens found by the \link{openNLP} tokenizer tool. These are
not always identical to the tokens one would find by running a simpler
tokenizer tool, like \code{\link[tokenizers:tokenize_words]{tokenizers::tokenize_words()}}.}

\item{tags}{The part of speech (POS) tag, from the \href{https://www.ling.upenn.edu/courses/Fall_2003/ling001/penn_treebank_pos.html}{PennTreebank}.
These tags are generated by running each headline through
\code{\link[openNLP:Maxent_POS_Tag_Annotator]{openNLP::Maxent_POS_Tag_Annotator()}}. More accurate results could perhaps be
obtained by running the headlines through the a state-of-the art parser like
Stanford POS annotator (see http://stanfordnlp.github.io/CoreNLP/). But that
requires more memory and processing power than I had!}

\item{probs}{The estimated probability (by
\code{\link[openNLP:Maxent_POS_Tag_Annotator]{openNLP::Maxent_POS_Tag_Annotator()}}) that the word actually functions as
the tag says it does.}

}
}
\seealso{
Other headline data: \code{\link{headlines}},
  \code{\link{preprocessed_corpus}}
}
\keyword{datasets}
